+++ app/src/main/java/com/yeahdev/yeahstreamer/models/RadioStation.java
@@ -1,4 +1,4 @@
-package com.yeahdev.yeahstreamer.model;
+package com.yeahdev.yeahstreamer.models;
 
 
 public class RadioStation {
