+++ app/src/main/java/com/yeahdev/yeahstreamer/utils/Constants.java
@@ -1,4 +1,4 @@
-package com.yeahdev.yeahstreamer.util;
+package com.yeahdev.yeahstreamer.utils;
 
 public class Constants {
